requirements.txt
asgiref==3.5.2
coverage==6.5.0
Django==4.1.4
django-cors-headers==3.13.0
django-jazzmin==2.6.0
django-material-admin==1.8.6
djangorestframework==3.14.0
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
asgiref==3.5.2
coverage==6.5.0
Django==4.1.4
django-cors-headers==3.13.0
django-jazzmin==2.6.0
django-material-admin==1.8.6
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
